backend/requirements.txt
fastapi~=0.115.12
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
Authlib==1.4.1
beautifulsoup4==4.12.2
build==1.2.2
CacheControl==0.14.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
colorama==0.4.6
crashtest==0.4.1
cryptography==44.0.1
deepseek==0.0.1.dev3
Deprecated==1.2.18
distlib==0.3.8
distro==1.9.0
dulwich==0.21.7
ecdsa==0.19.0
fastapi-auth0==0.5.0
fastjsonschema==2.20.0
filelock==3.16.1
firebase-admin==6.6.0
fuzzywuzzy==0.18.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
idna==3.10
importlib-resources==6.1.1
installer==0.7.0
jaraco.classes==3.4.0
jiter==0.8.2
keyring==24.3.1
limits==4.0.1
lxml==4.9.3
more-itertools==10.5.0
msgpack==1.1.0
openai==1.62.0
outcome==1.3.0.post0
packaging==24.2
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.3.6
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
proto-plus==1.26.0
protobuf==5.29.3
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.1.0
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dotenv==1.0.1
python-jose==3.3.0
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.10.0
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
selenium==4.26.1
shellingham==1.5.4
shin==0.2.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.45.3
tomlkit==0.13.2
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
trove-classifiers==2024.9.12
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.26.6
watchfiles==1.0.4
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
wsproto==1.2.0
